Move lib files to lib directory.
