+++ lib/hid/hid.c
