+++ lib/rscode/berlekamp.c
