+++ lib/rscode/crcgen.c
