+++ lib/rscode/galois.c
