+++ lib/rscode/rs.c
