+++ lib/rscode/rs.doc
